LayoutTests/platform/mac/accessibility/string-range-contains-listmarker-expected.txt

test
This tests makes sure that the list marker text is included when the text is asked for.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS succeeded is true
PASS successfullyParsed is true

TEST COMPLETE
